pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markpigeon"
version = "1.0.0"
description = "A powerful Markdown to HTML converter with smart asset management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Steven Li", email = "steven.jianhao.li@example.com"}
]
keywords = ["markdown", "html", "converter", "typora", "theme"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]
requires-python = ">=3.10"
dependencies = [
    "markdown-it-py>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "Pillow>=10.0.0",
    "PySide6>=6.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
markpigeon = "src.interfaces.cli.main:main"

[project.urls]
Homepage = "https://github.com/steven-jianhao-li/MarkPigeon"
Repository = "https://github.com/steven-jianhao-li/MarkPigeon"
Issues = "https://github.com/steven-jianhao-li/MarkPigeon/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
